Unit tests left queries and validators for books

+++ Library.Application/BookUseCases/Commands/DeleteBookHandler.cs
@@ -12,7 +12,10 @@ namespace Library.Application.BookUseCases.Commands
         public async Task Handle(DeleteBookCommand request, CancellationToken cancellationToken)
         {
             var book = await _unitOfWork.BookRepository.GetByIdAsync(request.Id);
-            
+
+            if (!book.IsAvailable)
+                throw new BookInUseException(book.Id);
+
             _unitOfWork.BookRepository.Delete(book);
             await _unitOfWork.SaveChangesAsync();
         }
